🔨 refactor: reorganize logger
